+++ example/oglplus/019_tessellation.cpp
@@ -314,7 +314,7 @@ public:
 		//
 		auto camera = CamMatrixf::Orbiting(
 			Vec3f(),
-			14.0 - SineWave(time / 13)*8.0,
+			GLfloat(14.0 - SineWave(time / 13)*8.0),
 			Degrees(time * 33),
 			Degrees(SineWave(time / 21.0) * 31)
 		);
